+++ pakyow-core/spec/features/commands/info_spec.rb
@@ -18,7 +18,9 @@ RSpec.describe "cli: info" do
 
   describe "help" do
     it "is helpful" do
-      expect(run_command(command, help: true, project: true)).to eq("\e[34;1mShow details about the current project\e[0m\n\n\e[1mUSAGE\e[0m\n  $ pakyow info\n\n\e[1mOPTIONS\e[0m\n  -e, --env=env  \e[33mThe environment to run this command under\e[0m\n")
+      cached_expectation "commands/info/help" do
+        run_command(command, help: true, project: true)
+      end
     end
   end