+++ pakyow-core/spec/features/commands/prelaunch/build_spec.rb
@@ -10,7 +10,9 @@ RSpec.describe "cli: prelaunch:build" do
 
   describe "help" do
     it "is helpful" do
-      expect(run_command(command, help: true, project: true, tty: false)).to eq("\e[34;1mRun the build phase of the prelaunch sequence\e[0m\n\n\e[1mUSAGE\e[0m\n  $ pakyow prelaunch:build\n\n\e[1mOPTIONS\e[0m\n  -e, --env=env  \e[33mThe environment to run this command under\e[0m\n")
+      cached_expectation "commands/prelaunch/build/help" do
+        run_command(command, help: true, project: true, tty: false)
+      end
     end
   end