+++ pakyow-data/spec/features/commands/bootstrap_spec.rb
@@ -14,7 +14,9 @@ RSpec.describe "cli: db:bootstrap" do
 
   describe "help" do
     it "is helpful" do
-      expect(run_command(command, help: true, project: true)).to eq("\e[34;1mBootstrap a database\e[0m\n\n\e[1mUSAGE\e[0m\n  $ pakyow db:bootstrap\n\n\e[1mOPTIONS\e[0m\n  -e, --env=env                \e[33mThe environment to run this command under\e[0m\n      --adapter=adapter        \e[33mThe database adapter (default: sql)\e[0m\n  -c, --connection=connection  \e[33mThe database connection (default: default)\e[0m\n")
+      cached_expectation "commands/bootstrap/help" do
+        run_command(command, help: true, project: true)
+      end
     end
   end